Renames implemnations => config linters example
