+++ phpbench.json
@@ -16,12 +16,11 @@
     "runner.revs": 10,
     "runner.assert": "mode(variant.time.avg) as ms <= mode(baseline.time.avg) as ms +/- 5%",
     "report.generators": {
-        "all": {
+        "examples": {
             "generator": "composite",
             "reports": [
-                "env",
-                "default",
-                "aggregate"
+                "hashing",
+                "config_linters"
             ]
         }
     },
